+++ res/examples/spawn.lua
@@ -1,0 +1,3 @@
+-- Spawns group ID 1
+
+Object.spawn(1, 2) -- group 1 with delay 2 seconds